test(item-sliding): migrate basic
